mqtt_subscriber/requirements.txt

paho-mqtt
confluent-kafka
